refactor the game rules and game initialization to simplify files for readability
