+++ app/game_controls.py
@@ -2,7 +2,6 @@ import sys
 
 import pygame
 
-import app.game_functions as gf
 import app.game_initialization as gi
 
 
@@ -30,7 +29,7 @@ def check_keydown_events(event, ai_settings, stats, screen, ship, bullets):
     elif event.key == pygame.K_LEFT:
         ship.moving_left = True
     elif event.key == pygame.K_SPACE:
-        gf.fire_bullet(ai_settings, bullets, screen, ship)
+        fire_bullet(ai_settings, bullets, screen, ship)
     elif event.key == pygame.K_p:
         stats.game_active = True
     elif event.key == pygame.K_q:
@@ -42,3 +41,9 @@ def check_keyup_events(event, ship):
         ship.moving_right = False
     elif event.key == pygame.K_LEFT:
         ship.moving_left = False
+
+
+def fire_bullet(ai_settings, bullets, screen, ship):
+    if len(bullets) < ai_settings.bullets_allowed:
+        new_bullet = Bullet(ai_settings, screen, ship)
+        bullets.add(new_bullet)
