+++ app/game_rules.py
@@ -1,30 +1,11 @@
-import sys
 from time import sleep
 
 import pygame
 
-from app.bullet import Bullet
-from app.alien import Alien
+import app.game_initialization as gi
 
 
-# Bullet handling
-def fire_bullet(ai_settings, bullets, screen, ship):
-    if len(bullets) < ai_settings.bullets_allowed:
-        new_bullet = Bullet(ai_settings, screen, ship)
-        bullets.add(new_bullet)
-
-
-def update_bullets(ai_settings, screen, stats, sb, ship, aliens, bullets):
-    """Update position of bullets and get rid of old bullets"""
-    bullets.update()
-
-    for bullet in bullets.copy():
-        if bullet.rect.bottom <= 0:
-            bullets.remove(bullet)
-
-    check_bullet_alien_collisions(ai_settings, screen, stats, sb, ship, aliens, bullets)
-
-
+# Game Rules
 def check_bullet_alien_collisions(ai_settings, screen, stats, sb, ship, aliens, bullets):
     """Check if bullet hit alien and increase level, speed, and score when fleet is destroyed"""
     collisions = pygame.sprite.groupcollide(bullets, aliens, True, True)
@@ -37,16 +18,18 @@ def check_bullet_alien_collisions(ai_settings, screen, stats, sb, ship, aliens, 
 
 
 def increase_score(ai_settings, aliens, sb, stats):
+    """Increase score when alien is shot"""
     stats.score += ai_settings.alien_points * len(aliens)
     sb.prep_score()
 
 
 def start_new_level(ai_settings, aliens, bullets, sb, screen, ship, stats):
+    """Start new level when fleet is destroyed"""
     bullets.empty()
     ai_settings.increase_speed()
     stats.level += 1
     sb.prep_level()
-    create_fleet(ai_settings, screen, ship, aliens)
+    gi.create_fleet(ai_settings, screen, ship, aliens)
 
 
 def check_high_score(stats, sb):
@@ -54,49 +37,6 @@ def check_high_score(stats, sb):
     if stats.score > stats.high_score:
         stats.high_score = stats.score
         sb.prep_high_score()
-
-
-# Alien handling
-def create_fleet(ai_settings, screen, ship, aliens):
-    """Create a full fleet of aliens"""
-
-    alien = Alien(ai_settings, screen)
-    number_aliens_x = get_number_aliens_x(ai_settings, alien.rect.width)
-    number_rows = get_number_rows(ai_settings, ship.rect.height, alien.rect.height)
-
-    # Create the first row of aliens
-    for row_number in range(number_rows):
-        for alien_number in range(number_aliens_x):
-            create_alien(ai_settings, screen, aliens, alien_number, row_number)
-
-
-def create_alien(ai_settings, screen, aliens, alien_number, row_number):
-    alien = Alien(ai_settings, screen)
-    alien_width = alien.rect.width
-    alien.x = alien_width + ai_settings.alien_space * alien_width * alien_number
-    alien.rect.x = alien.x
-    alien.rect.y = alien.rect.height + 2 * alien.rect.height * row_number
-    aliens.add(alien)
-
-
-def get_number_aliens_x(ai_settings, alien_width):
-    available_space_x = ai_settings.screen_width - 5 * alien_width
-    number_aliens_x = int(available_space_x / (ai_settings.alien_space * alien_width))
-    return number_aliens_x
-
-
-def get_number_rows(ai_settings, ship_height, alien_height):
-    available_space_y = (ai_settings.screen_height - (4 * alien_height) - ship_height)
-    number_rows = int(available_space_y / (3 * alien_height))
-    return number_rows
-
-
-def check_fleet_edges(ai_settings, aliens):
-    """Respond appropriately if any aliens have reached the edge"""
-    for alien in aliens.sprites():
-        if alien.check_edges():
-            change_fleet_direction(ai_settings, aliens)
-            break
 
 
 def change_fleet_direction(ai_settings, aliens):
@@ -111,13 +51,19 @@ def ship_hit(ai_settings, stats, screen, sb, ship, aliens, bullets):
     if stats.ships_left > 0:
         stats.ships_left -= 1
         sb.prep_ships()
-        aliens.empty()
-        bullets.empty()
-        create_fleet(ai_settings, screen, ship, aliens)
-        ship.center_ship()
+        gi.reset_ship(bullets, ship)
+        gi.reset_fleet(ai_settings, aliens, screen, ship)
         sleep(0.5)
     else:
         stats.game_active = False
+
+
+def check_fleet_edges(ai_settings, aliens):
+    """Respond appropriately if any aliens have reached the edge"""
+    for alien in aliens.sprites():
+        if alien.check_edges():
+            change_fleet_direction(ai_settings, aliens)
+            break
 
 
 def check_aliens_bottom(ai_settings, stats, screen, sb, ship, aliens, bullets):
@@ -129,8 +75,19 @@ def check_aliens_bottom(ai_settings, stats, screen, sb, ship, aliens, bullets):
             break
 
 
+def update_bullets(ai_settings, screen, stats, sb, ship, aliens, bullets):
+    """Update position of bullets and get rid of old bullets"""
+    bullets.update()
+
+    for bullet in bullets.copy():
+        if bullet.rect.bottom <= 0:
+            bullets.remove(bullet)
+
+    check_bullet_alien_collisions(ai_settings, screen, stats, sb, ship, aliens, bullets)
+
+
 def update_aliens(ai_settings, stats, screen, sb, ship, aliens, bullets):
-    """Check uf the fleet is at an edge, and then update the positions of all aliens in the fleet"""
+    """Check if the fleet is at an edge, and then update the positions of all aliens in the fleet"""
     check_fleet_edges(ai_settings, aliens)
     check_aliens_bottom(ai_settings, stats, screen, sb, ship, aliens, bullets)
     aliens.update()
